Primera cotización prueba del ctizador api
