+++ public/plantilla/js/config/api.js
@@ -8,9 +8,22 @@ const pathEstadisticasCiudad = pathCiudades + "/Estadisticas";
 
 const pathCotizador = endpoint + "/Heka/Cotizar";
 
-export {
+const v0 = {
     pathCiudadesLista,
     pathCiudadDane,
     pathEstadisticasCiudad,
     pathCotizador
+}
+
+const v1 = {
+    quoter: "https://api.hekaentrega.co/Api" + "/v1/shipping/quoter"
+}
+
+export {
+    pathCiudadesLista,
+    pathCiudadDane,
+    pathEstadisticasCiudad,
+    pathCotizador,
+    v0,
+    v1
 }
